Move sarif parsing code to a location that can be shared.
